Apologies for the large commit--fixing the pipes got out of hand.
In this code, I rework the parallel code some more, because it was not consistently working as intended. I think my current approach fixes things by reconfiguring declare_population and the sim_meta function. The when() function did not play nice with native pipes.

Near the end of the code, I also had to do some reconfiguring to get the native pipes to work.